+++ changelog.md
@@ -5,3 +5,5 @@
 - Added changelog - 16:10
 - Created develop branch - 16:10
 - Created pre-dev branch - 16:36
+- home.js 'About Me' button wrapped in class and styled - 16:47
+- added css box-sizing to project-tile components - 17:03